FEAT: DISEÑO DE PUNTUACIONES
El contenedor de las puntuaciones ha comenzado a ser diseñado
