+++ client/src/pages/publicProfile.js
@@ -1,16 +1,17 @@
 import '../static/css/pages/publicProfile.css';
 
 import FlatterPage from "../sections/flatterPage";
-import CommentsBox from "../sections/commentsBox";
+import CommentsBox from "../components/profile/commentsBox";
 import PublicProfileCard from "../components/profile/publicProfileCard";
+import ReviewsBox from "../components/profile/reviewsBox";
 
 const PublicProfile = () => {
 
     return(
-        <FlatterPage withBackground userLogged>
+        <FlatterPage withBackground>
             <div className="profile-grid">
                 <PublicProfileCard name="Lucía Martín" job="Estudiante en Universidad de Sevilla" age={24} bio="Soy una estudiante a la que le pasárselo bien pero que también es responsable." isMe={true} isTenant={true}/>
-                <span>Pruebas</span>
+                <ReviewsBox average={5} total={0} />
                 <CommentsBox />
             </div>
         </FlatterPage>
