+++ client/src/static/css/components/commentsBox.css
@@ -48,12 +48,23 @@
     outline: none;
 }
 
+.comments-box-star{
+    width: 25px;
+    height: 25px;
+}
+
 .profile-grid{
     display: grid;
     grid-template-columns: 75% 25%;
     row-gap: 50px;
     column-gap: 30px;
     margin-top: 3em;
+}
+
+.comments-box-title{
+    margin-left: 1.5em;
+    font-weight: bold;
+    text-transform: uppercase;
 }
 
 @media screen and (max-width: 1319px) {
